+++ ignite/cmd/scaffold_chain.go
@@ -98,7 +98,10 @@ func scaffoldChainHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	appdir, err := scaffolder.Init(cacheStorage, placeholder.New(), appPath, name, addressPrefix, noDefaultModule)
+	appdir, err := scaffolder.Init(
+		cmd.Context(), cacheStorage, placeholder.New(), appPath, name,
+		addressPrefix, noDefaultModule,
+	)
 	if err != nil {
 		return err
 	}