+++ ignite/pkg/cosmoscmd/cosmoscmd.go
@@ -1,0 +1,2 @@
+// Deprecated: the content of this package is now scaffolded in chains.
+package cosmoscmd